Use helper functions, move workflow-only functions into script
