+++ e2e-test/cromwell_azure_e2e_test.py
@@ -8,7 +8,9 @@ import time
 from collections import deque
 import sys
 sys.path.append(os.path.join(os.path.dirname(__file__), '..', '..'))
-from helper import get_completed_workflow, submit_hello_world_to_cromwell, delete_workspace, output_message, handle_failed_request
+from helper import output_message, handle_failed_request
+from workspace_helper import create_workspace, delete_workspace
+from app_helper import create_app, poll_for_app_url
 
 bearer_token = os.environ['BEARER_TOKEN']
 bee_name = os.environ['BEE_NAME']
@@ -17,69 +19,74 @@ billing_project_name = os.environ['BILLING_PROJECT_NAME']
 rawls_url = f"https://rawls.{bee_name}.bee.envs-terra.bio"
 leo_url = f"https://leonardo.{bee_name}.bee.envs-terra.bio"
 
-def create_workspace():
-   rawls_api_call = f"{rawls_url}/api/workspaces"
-   request_body= {
-      "namespace": billing_project_name, # Billing project name
-      "name": f"api-workspace-{''.join(random.choices(string.ascii_lowercase, k=5))}", # workspace name
-      "attributes": {}}
-   
-   create_workspace_response = requests.post(url=rawls_api_call, 
-                                             json=request_body, 
-                                             headers={"Authorization": f"Bearer {bearer_token}"}
-   ).json()
-
-   create_workspace_data = json.loads(json.dumps(create_workspace_response))
-   workspace_id = create_workspace_data['workspaceId']
-
-   output_message(f"Enabling Cromwell for workspace {workspace_id}")
-   activate_cromwell_request = f"{leo_url}/api/apps/v2/{workspace_id}/terra-app-{str(uuid.uuid4())}"
-   cromwell_request_body = {
-      "appType": "CROMWELL_RUNNER"
-   } 
+def submit_hello_world_to_cromwell(app_url, workflow_test_name):
+    absolute_file_path = os.path.dirname(__file__)
+    workflow_source_path = os.path.join(absolute_file_path, './resources/cromwell/hello.wdl')
+    workflow_inputs_path = os.path.join(absolute_file_path, './resources/cromwell/hello.inputs')
+    workflow_endpoint = f'{app_url}/api/workflows/v1'
+    headers = {"Authorization": f'Bearer {bearer_token}',
+              "accept": "application/json",
+    }
+    with open(workflow_source_path, 'rb') as hello_wdl:
+        with open(workflow_inputs_path, 'rb') as hello_inputs:
+            files = {
+                'workflowSource': ('hello.wdl', hello_wdl, 'application/octet-stream'),
+                'workflowInputs': ('hello.inputs', hello_inputs, 'application/octet-stream'),
+                'workflowType': 'WDL',
+                'workflowTypeVersion': '1.0'
+            }
+            response = requests.post(workflow_endpoint, headers=headers, files=files)
+            handle_failed_request(response, f"Error submitting workflow to Cromwell for {workflow_test_name}", 201)
+            output_message(response.json(), "DEBUG")
+            return response.json()
         
-   response = requests.post(url=activate_cromwell_request, json=cromwell_request_body, 
-                            headers={"Authorization": f"Bearer {bearer_token}"})
-   # will return 202 or error
-   handle_failed_request(response, "Error activating Cromwell", 202)
-   output_message("Cromwell successfully activated")
-   return create_workspace_data
+def get_workflow_information(app_url, workflow_id, bearer_token):
+    workflow_endpoint = f'{app_url}/api/workflows/v1/{workflow_id}/status'
+    headers = {"Authorization": f'Bearer {bearer_token}',
+              "accept": "application/json"}
+    output_message(f"Fetching workflow status for {workflow_id}")
+    response = requests.get(workflow_endpoint, headers=headers)
+    handle_failed_request(response, f"Error fetching workflow metadata for {workflow_id}")
+    output_message(response.json(), "DEBUG")
+    return response.json()
 
-# GET CROMWELL ENDPOINT URL FROM LEO
-def get_app_url(workspaceId, app):
-    uri = f"{leo_url}/api/apps/v2/{workspaceId}?includeDeleted=false"
+# workflow_ids is a deque of workflow ids
+def get_completed_workflow(app_url, workflow_ids, max_retries=4, sleep_timer=60 * 2):
+    success_statuses = ['Succeeded']
+    throw_exception_statuses = ['Aborted', 'Failed']
+    
+    current_running_workflow_count = 0
+    while workflow_ids:
+        if max_retries == 0:
+            raise Exception(f"Workflow(s) did not finish running within retry window ({max_retries} retries)")
+        
+        workflow_id = workflow_ids.pop()
+        workflow_metadata = get_workflow_information(app_url, workflow_id, bearer_token)
+        workflow_status = workflow_metadata['status']
 
-    headers = {"Authorization": f"Bearer {bearer_token}",
-               "accept": "application/json"}
-
-    response = requests.get(uri, headers=headers)
-    status_code = response.status_code
-    handle_failed_request(response, f"{status_code} - Error retrieving details for {workspaceId}", 200)
-    output_message(f"Successfully retrieved details.\n{response.text}")
-    response = response.json()
-
-    app_url = ""
-    app_type = "CROMWELL_RUNNER" if app != 'wds' else app.upper()
-    output_message(f"App type: {app_type}")
-    for entries in response: 
-        if entries['appType'] == app_type and entries['proxyUrls'][app] is not None:
-            if(entries['status'] == "PROVISIONING"):
-                output_message(f"{app} is still provisioning")
-                break
-            output_message(f"App status: {entries['status']}")
-            app_url = entries['proxyUrls'][app]
-            break 
-
-    if app_url is None: 
-        output_message(f"{app} is missing in current workspace")
-    else:
-        output_message(f"{app} url: {app_url}")
-
-    return app_url
+        if(workflow_status in throw_exception_statuses):
+            raise Exception(f"Exception raised: Workflow {workflow_id} reporting {workflow_status} status")
+        if workflow_status in success_statuses:
+            output_message(f"{workflow_id} finished running. Status: {workflow_metadata['status']}")
+        else:
+            workflow_ids.appendleft(workflow_id)
+            current_running_workflow_count += 1
+        if current_running_workflow_count == len(workflow_ids):
+            if current_running_workflow_count == 0:
+                output_message("Workflow(s) finished running")
+            else:
+                # Reset current count to 0 for next retry
+                # Decrement max_retries by 1
+                # Wait for sleep_timer before checking workflow statuses again (adjust value as needed)
+                output_message(f"These workflows have yet to return a completed status: [{', '.join(workflow_ids)}]")
+                max_retries -= 1
+                current_running_workflow_count = 0
+                time.sleep(sleep_timer)
+    output_message("Workflow(s) submission and completion successful")
         
 def test_cleanup(workspace_namespace, workspace_name):
     try:
-        delete_workspace(workspace_namespace, workspace_name)
+        delete_workspace(workspace_namespace, workspace_name, rawls_url, bearer_token)
         output_message("Workspace cleanup complete")
     # Catch the exeception and continue with the test since we don't want cleanup to affect the test results
     # We can assume that Janitor will clean up the workspace if the test fails
@@ -88,23 +95,20 @@ def test_cleanup(workspace_namespace, workspace_name):
         output_message(f'Exception details below:\n{e}')
 
 def main():
-    workspace_namespace = ""
     workspace_name = ""
     workspace_id = ""
     found_exception = False
     
     # Sleep timers for various steps in the test
     workflow_run_sleep_timer = 60 * 5
-    provision_sleep_timer = 60 * 15
+    provision_sleep_timer = 60 * 5
     workflow_status_sleep_timer = 60 * 2
     
     try:
-        created_workspace = create_workspace()
-        workspace_id = created_workspace['workspaceId']
-        workspace_namespace = created_workspace['namespace']
-        workspace_name = created_workspace['name']
-        time.sleep(provision_sleep_timer) # Added an sleep here to give the workspace time to provision
-        app_url = get_app_url(workspace_id, 'cromwell')
+        (workspace_id, workspace_name) = create_workspace(billing_project_name, bearer_token, rawls_url)
+        create_app(workspace_id, leo_url, 'CROMWELL_RUNNER', 'USER_PRIVATE', bearer_token)
+        time.sleep(provision_sleep_timer) # Added an sleep here to give the workspace time to provision and app to start
+        app_url = poll_for_app_url(workspace_id, 'CROMWELL_RUNNER', 'cromwell-runner', bearer_token, leo_url)
 
         # This chunk of code only executes one workflow
         # Would like to modify this down the road to execute and store references for multiple workflows
@@ -115,12 +119,12 @@ def main():
         # This chunk of code supports checking one or more workflows
         # Probably won't require too much modification if we want to run additional submission tests
         workflow_ids = deque([workflow_response['id']])
-        get_completed_workflow(app_url, workflow_ids, workflow_status_sleep_timer)
+        get_completed_workflow(app_url, workflow_ids, sleep_timer=workflow_status_sleep_timer)
     except Exception as e:
         output_message(f"Exception occured during test:\n{e}")
         found_exception = True
     finally:
-        test_cleanup(workspace_namespace, workspace_name)
+        test_cleanup(billing_project_name, workspace_name)
         # Use exit(1) so that GHA will fail if an exception was found during the test
         if(found_exception):
             output_message("Workflow test failed due to exception(s)", "ERROR")
